requirements.txt
Django>=5.1.4
djangorestframework>=3.14.0
psycopg2-binary>=2.9.10
drf-spectacular>=0.27.0 